pyproject.toml
[project]
name = "dbx-fetch-benchmark"
version = "0.0.0"
description = "Databricks fetch benchmark runners"
requires-python = ">=3.10"
dependencies = [
  "delta-kernel-rust-sharing-wrapper>=0.3.1",
  "delta-sharing>=1.3.0",
  "adbc-driver-manager>=1.9.0",
  "databricks-sdk>=0.72.0",
  "databricks-sql-connector>=3.0.0",
  "duckdb>=1.4.1",
  "jaydebeapi>=1.2.3",
  "JPype1>=1.5.0",
  "pandas>=2.0.0",
  "pyarrow>=18.0.0",
  "pyodbc>=5.0.0",
  "python-dotenv>=1.0.0",
]
